@@ -1,58 +1,56 @@
 <template>
-  <div id="app">
-    <router-view :activeUser="activeUser"/>
-    <footer class="info">
-      <p v-if="activeUser" class="logout-link"><a @click="handleLogout" href="#">Logout</a></p>
-      <p>Based on a project written by <a href="http://evanyou.me">Evan You</a></p>
-      <p>Original Vue TodoApp project is <a href="https://vuejs.org/v2/examples/todomvc.html">here</a></p>
-      <p>Modified for this tutorial by Andrew Hughes</p>
-    </footer>
-  </div>
+  <q-layout view="hHh lpR fFf">
+
+    <q-header elevated class="bg-primary text-white">
+      <q-toolbar>
+        <q-toolbar-title>
+          <q-avatar>
+            <q-icon name="kayaking" size="30px"></q-icon>
+          </q-avatar>
+          Todo App
+        </q-toolbar-title>
+        {{ this.claims && this.claims.email ? claims.email : '' }}
+        <q-btn flat round dense icon="logout" v-if='authState && authState.isAuthenticated' @click="logout"/>
+        <q-btn flat round dense icon="account_circle" v-else @click="login"/>
+      </q-toolbar>
+    </q-header>
+
+    <q-page-container>
+      <router-view></router-view>
+    </q-page-container>
+
+  </q-layout>
 </template>
 
 <script>
-
-  // app Vue instance
-  const app = {
-
-    name: 'app',
-
-    // app initial state
-    data: () => {
-      return {
-        activeUser: null
+export default {
+  name: 'LayoutDefault',
+  data: function () {
+    return {
+      claims: null
+    }
+  },
+  watch: {
+    'authState.isAuthenticated'() {
+      this.$log.debug(("watch triggered!"))
+      this.updateClaims()
+    }
+  },
+  created() {
+    this.updateClaims()
+  },
+  methods: {
+    async updateClaims() {
+      if (this.authState && this.authState.isAuthenticated) {
+        this.claims = await this.$auth.getUser()
       }
     },
-
-    async created () {
-      await this.refreshActiveUser()
+    async login() {
+      await this.$auth.signInWithRedirect({ originalUri: '/todos' })
     },
-
-    watch: {
-      '$route': 'refreshActiveUser'
-    },
-
-    methods: {
-      async refreshActiveUser () {
-        this.activeUser = await this.$auth.getUser()
-        this.$log.debug('activeUser',this.activeUser)
-      },
-
-      async handleLogout () {
-        await this.$auth.logout()
-        await this.refreshActiveUser()
-        this.$router.go('/')
-      }
-    },
-
-  }
-
-  export default app
-
+    async logout() {
+      await this.$auth.signOut()
+    }
+  },
+}
 </script>
-
-<style>
-
-  [v-cloak] { display: none; }
-
-</style>
